requirements.txt
requests
bs4
lxml
python-daemon-3K
defusedxml